refactor(setup): extract wizard step 6 (hosts modification)
